# javalin-maven-kotlin-basic/src/main/resources/public/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "",
    "version" : ""
  },
  "paths" : {
    "/hello" : {
      "post" : {
        "summary" : "",
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HiDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "No content"
          }
        }
      }
    },
    "/hello/:id" : {
      "get" : {
        "summary" : "Find the Hello by id",
        "description" : "This has some more interesting detail to document.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The hello identity",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The hello",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HiDto"
                }
              }
            }
          }
        }
      }
    },
    "/hello/asform" : {
      "post" : {
        "summary" : "",
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/HelForm"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "No content"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "HelForm" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "email" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "HiDto" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "other" : {
            "type" : "string"
          }
        }
      }
    }
  }
}